menu - reorganize into files, simplify first&second level, document
