+++ app/javascript/menu/menu-collapse.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { ChevronLeft20, ChevronRight20 } from '@carbon/icons-react';
+import { SideNavItem } from 'carbon-components-react/es/components/UIShell';
+
+export const MenuCollapse = ({ expanded, toggle }) => (
+  <SideNavItem className="menu-collapse">
+    <div
+      className="menu-collapse-button"
+      onClick={toggle}
+    >
+      {expanded ? <ChevronLeft20 /> : <ChevronRight20 />}
+    </div>
+  </SideNavItem>
+);
